requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.12.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.2.0
timm>=0.9.0
torchsummary>=1.5.1
pillow>=9.5.0
pandas>=2.0.0